refactor: @Transactional 어노테이션 리팩토링
