+++ src/main/java/com/example/pace/domain/auth/service/command/AuthCommandService.java
@@ -8,79 +8,53 @@ import com.example.pace.domain.auth.service.query.KakaoApiQueryService;
 import com.example.pace.domain.member.dto.response.KakaoUserInfoResDTO;
 import com.example.pace.domain.member.entity.Member;
 import com.example.pace.domain.member.enums.Role;
-import com.example.pace.domain.member.enums.SocialProvider;
 import com.example.pace.domain.member.exception.code.MemberErrorCode;
 import com.example.pace.domain.member.exception.MemberException;
 import com.example.pace.domain.member.repository.MemberRepository;
+import com.example.pace.domain.member.service.command.MemberCommandService;
 import com.example.pace.global.util.JwtUtil;
 import io.jsonwebtoken.Claims;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-@Transactional
 public class AuthCommandService {
     private final MemberRepository memberRepository;
     private final KakaoApiQueryService kakaoApiQueryService;
+    private final MemberCommandService memberCommandService;
     private final JwtUtil jwtUtil;
 
     public AuthResDTO.LoginResultDTO loginWithKakao(String kakaoAccessToken) {
         // 카카오 API로 사용자 정보 조회
         KakaoUserInfoResDTO kakaoUserInfoResDTO = kakaoApiQueryService.getUserInfo(kakaoAccessToken);
-        String email = kakaoUserInfoResDTO.getKakaoAccount().getEmail();
-        String socialId = kakaoUserInfoResDTO.getId().toString();
 
-        Optional<Member> memberOptional = memberRepository.findBySocialProviderAndSocialIdIgnoreStatus(
-                SocialProvider.KAKAO.name(),
-                socialId
-        );
+        Member member = memberCommandService.getOrSaveMember(kakaoUserInfoResDTO);
 
-        // 값이 있을 경우(기존 회원일 경우)
-        if (memberOptional.isPresent()) {
-            Member member = memberOptional.get();
-
-            // 온보딩 미완료 체크
-            if (member.getRole() == Role.ROLE_INCOMPLETE_USER) {
-                String tempToken = jwtUtil.createTempToken(member.getId());
-
-                // isNewUser 값을 true인 채로 반환
-                return AuthConverter.toNewMemberDTO(member, tempToken);
-            }
-
-            // 조회된 회원이 탈퇴한 상태인지 확인
-            if (!member.getIsActive()) {
-                throw new MemberException(MemberErrorCode.MEMBER_NOT_ACTIVE);
-            }
-
-            String accessToken = jwtUtil.createAccessToken(member.getId(), Role.ROLE_USER);
-            String refreshToken = jwtUtil.createRefreshToken(member.getId());
-            member.updateRefreshToken(refreshToken);
-
-            return AuthConverter.toExistingMemberDTO(member, accessToken, refreshToken);
-        } else {
-            // 신규 회원일 경우
-            Member newMember = Member.builder()
-                    .socialId(socialId)
-                    .email(email)
-                    .nickname(kakaoUserInfoResDTO.getKakaoAccount().getProfile().getNickname())
-                    .role(Role.ROLE_INCOMPLETE_USER)
-                    .socialProvider(SocialProvider.KAKAO)
-                    .isActive(true)
-                    .build();
-
-            memberRepository.save(newMember);
-
-            // 온보딩용 임시 토큰 발급(10분 유효)
-            String tempToken = jwtUtil.createTempToken(newMember.getId());
-
-            return AuthConverter.toNewMemberDTO(newMember, tempToken);
+        // 비활성된(탈퇴) 회원인 경우
+        if (!member.getIsActive()) {
+            throw new MemberException(MemberErrorCode.MEMBER_NOT_ACTIVE);
         }
+
+        // 온보딩 미완료 회원일 경우
+        if (member.getRole() == Role.ROLE_INCOMPLETE_USER) {
+            String tempToken = jwtUtil.createTempToken(member.getId());
+
+            return AuthConverter.toNewMemberDTO(member, tempToken);
+        }
+
+        String accessToken = jwtUtil.createAccessToken(member.getId(), member.getRole());
+        String refreshToken = jwtUtil.createRefreshToken(member.getId());
+
+        // 리프레쉬 토큰 업데이트
+        memberCommandService.updateRefreshToken(member.getId(), refreshToken);
+
+        return AuthConverter.toExistingMemberDTO(member, accessToken, refreshToken);
     }
 
     // 리프레쉬 토큰으로 액세스 토큰 재발행 로직
+    @Transactional
     public AuthResDTO.LoginResultDTO reissueToken(String refreshToken) {
         // 리프레쉬 토큰 유효성 검증
         if (!jwtUtil.validateToken(refreshToken)) {
