--- src/main/java/com/example/pace/domain/member/service/query/MemberQueryService.java
@@ -1,4 +1,0 @@
-package com.example.pace.domain.member.service.query;
-
-public class MemberQueryService {
-}
